+++ src/components/transformation_history.ts
@@ -3,11 +3,11 @@
 
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { DaCeInterface } from '../daceInterface';
+import { DaCeInterface } from '../dace_interface';
 import { DaCeVSCode } from '../extension';
 
-import { BaseComponent } from './baseComponent';
-import { ComponentMessageHandler } from './messaging/componentMessageHandler';
+import { BaseComponent } from './base_component';
+import { ComponentMessageHandler } from './messaging/component_message_handler';
 
 export class TransformationHistoryProvider
 extends BaseComponent
